Add feature | User authentication with JWT

+++ users/schema.py
@@ -1,8 +1,11 @@
 import graphene
 from graphene_django import DjangoObjectType
+from graphene_django.views import GraphQLError
+from django.db import connections
 from django.contrib.auth.hashers import make_password, check_password
 import uuid
 from users.models import User
+from users.utils import test_db_connection
 
 class UserType(DjangoObjectType):
   class Meta:
@@ -27,18 +30,21 @@ class CreateUser(graphene.Mutation):
   @staticmethod
   def mutate(self, info, user_input=None):
 
-    hashed_password = make_password(user_input.password)
+    if test_db_connection(connections):
+      hashed_password = make_password(user_input.password)
 
-    user_instance = User(
-      id = uuid.uuid4(),
-      username = user_input.username,
-      email = user_input.email,
-      bio = user_input.bio,
-      password = hashed_password
-    )
+      user_instance = User(
+        id = uuid.uuid4(),
+        username = user_input.username,
+        email = user_input.email,
+        bio = user_input.bio,
+        password = hashed_password
+      )
 
-    user_instance.save()
-    return CreateUser(user=user_instance)
+      user_instance.save()
+      return CreateUser(user=user_instance)
+    else:
+      raise GraphQLError('DB Server not connected!')
 
 class DeleteUser(graphene.Mutation):
   id = graphene.UUID()
@@ -49,10 +55,16 @@ class DeleteUser(graphene.Mutation):
   user = graphene.Field(UserType)
   @staticmethod
   def mutate(self, info, **kwargs):
-    user = User.objects.get(pk=kwargs.get('id'))
+    if test_db_connection(connections):
+      if not info.context.user.is_anonymous:
+        user = User.objects.get(pk=kwargs.get('id'))
 
-    user.delete()
-    return DeleteUser(deleted=True)
+        user.delete()
+        return DeleteUser(deleted=True)
+      else:
+        raise GraphQLError('You must be logged!')
+    else:
+      raise GraphQLError('DB Server not connected!')
 
 class UpdateUser(graphene.Mutation):
   
@@ -63,18 +75,25 @@ class UpdateUser(graphene.Mutation):
   user = graphene.Field(UserType)
   @staticmethod
   def mutate(self, info, user_input=None, **kwargs):
-    
-    user_instance = User.objects.get(pk=kwargs.get('id'))
-    if user_input.username:
-      user_instance.username = user_input.username
-    if user_input.email:
-      user_instance.email = user_input.email
-    if user_input.bio:
-      user_instance.bio = user_input.bio
-    if user_input.password:
-      user_instance.password = make_password(user_input.password)
+    if test_db_connection(connections):
+      if not info.context.user.is_anonymous:
+        user_instance = User.objects.get(pk=kwargs.get('id'))
+        if user_input.username:
+          user_instance.username = user_input.username
+        if user_input.email:
+          user_instance.email = user_input.email
+        if user_input.bio:
+          user_instance.bio = user_input.bio
+        if user_input.password:
+          user_instance.password = make_password(user_input.password)
 
-    return UpdateUser(user=user_instance)
+        user_instance.save()
+
+        return UpdateUser(user=user_instance)
+      else:
+        raise GraphQLError('You must be logged!')
+    else:
+      raise GraphQLError('DB Server not connected!')
 
 
 
@@ -83,10 +102,16 @@ class Query(graphene.ObjectType):
   user_by_id = graphene.Field(UserType, id=graphene.String())
 
   def resolve_users(self, info, **kwargs):
-    return User.objects.all()
+    if test_db_connection(connections):
+      return User.objects.all()
+    else:
+      raise GraphQLError('DB Server not connected!')
 
   def resolve_user_by_id(self, info, **kwargs):
-    return User.objects.get(pk=kwargs.get('id'))
+    if test_db_connection(connections):
+      return User.objects.get(pk=kwargs.get('id'))
+    else:
+      raise GraphQLError('DB Server not connected!')
 
 
 class Mutation(graphene.ObjectType):
